:sparkles: add support for per-profile container
The initial idea was to just open directly the
profile url in a new container but, if we do so,
we can't share the authentication session between
all the profile containers and one has to re-auth
for each profile container.

To be able to share the authentication session,
we now intercept the request used to obtain the
sign-in token. This request is done after successful
authentication has been performed in the sso
portal page, and the sign-in token allows us to build
a sign-in url that we can just pass to the new
container to successfully open the new profile in
another container.