:bug: fix breakage caused by new AWS Access Portal design
AWS deployed progressively a new design for their AWS SSO Portal Page UI
which breaks the parsing logic of this extension.
In addition, the format of the connection URL for each target profile
changed and doesn't include anymore the AWS account name information.

This commit fixes the issues by adding new parser logic for the new
layout. We still support the legacy AWS access portal as it might still
be available for some customers.

The new connection URL format is more tricky as the lack of account name
information will cause the created container to contain the account id
instead of the account name unless the user explicitly loaded the list of
profile name from the portal page first (which should be the usual usage
with this extension).

There is no perfect solution for that issue, for now we fallback to
displaying the account id instead of the account name.
We might improve that later if we can get the account name information
on the fly when the user went throught the SSO portal page before.