:recycle: move aws profile info in a dedicated class
By using a class, we can now have better flexibility to implement the
various attributes of an aws profile.
In particular, `title` and `id` are now dynamically computed base on
other attributes. It allows us to better support both the new and legacy
aws portal layout, as we can have some case where we don't have the
account name with the latter.